refactor(js): split monolith index.js into modules
