+++ fchat/apps/chat/templates/chat/room.html
@@ -64,7 +64,9 @@
   </div>
 </div>
 
-<script src="{% static 'js/room.js' %}"></script>
+  <script type="module" src="{% static 'js/chat_ui.js' %}"></script>
+  <script type="module" src="{% static 'js/connection.js' %}"></script>
+  <script type="module" src="{% static 'js/group_ui.js' %}"></script>
 <script src="{% static 'bootstrap/js/bootstrap.bundle.min.js' %}"></script>
 </body>
 </html>
