--- fchat/static/src/js/ui.js
@@ -1,1 +1,0 @@
-// раздроблю код из основного js
